pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hashzap"
version = "0.1.0"
description = "A professional real-time chat application"
authors = [
    {name = "Vitoria", email = "vitoria@example.com"},
]
dependencies = [
    "flask",
    "flask-socketio",
    "flet",
    "python-dotenv",
    "eventlet",
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
where = ["src"]
